@@ -1,58 +1,52 @@
 {
   "name": "averos-io-starter",
-  "version": "1.0.0",
+  "version": "0.0.0",
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
     "build": "ng build",
-    "test": "ng test",
-    "lint": "ng lint",
-    "e2e": "ng e2e"
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~12.0.4",
-    "@angular/cdk": "^12.0.4",
-    "@angular/common": "~12.0.4",
-    "@angular/compiler": "~12.0.4",
-    "@angular/core": "~12.0.4",
+    "@angular/animations": "~12.2.0",
+    "@angular/cdk": "^12.2.0",
+    "@angular/common": "~12.2.0",
+    "@angular/compiler": "~12.2.0",
+    "@angular/core": "~12.2.0",
     "@angular/flex-layout": "^12.0.0-beta.34",
-    "@angular/forms": "~12.0.4",
-    "@angular/material": "^12.0.4",
-    "@angular/platform-browser": "~12.0.4",
-    "@angular/platform-browser-dynamic": "~12.0.4",
-    "@angular/router": "~12.0.4",
-    "@ngrx/effects": "^12.1.0",
-    "@ngrx/entity": "^12.1.0",
-    "@ngrx/router-store": "^12.1.0",
-    "@ngrx/store": "^12.1.0",
-    "@ngrx/store-devtools": "^12.1.0",
-    "@wiforge/averos": "^1.1.6",
-    "file-saver": "^2.0.5",
+    "@angular/forms": "~12.2.0",
+    "@angular/material": "^12.2.0",
+    "@angular/platform-browser": "~12.2.0",
+    "@angular/platform-browser-dynamic": "~12.2.0",
+    "@angular/router": "~12.2.0",
+    "@ngrx/effects": "^12.3.0",
+    "@ngrx/entity": "^12.3.0",
+    "@ngrx/router-store": "^12.3.0",
+    "@ngrx/schematics": "^12.3.0",
+    "@ngrx/store": "^12.3.0",
+    "@ngrx/store-devtools": "^12.3.0",
+    "@wiforge/averos": "^1.2.1",
+    "file-saver": "^2.0.2",
     "rxjs": "~6.6.0",
-    "tslib": "^2.0.0",
-    "xlsx": "^0.16.9",
+    "tslib": "^2.3.0",
+    "xlsx": "^0.16.7",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~12.0.4",
-    "@angular/cli": "~12.0.4",
-    "@angular/compiler-cli": "~12.0.4",
-    "@angular/localize": "^12.0.4",
-    "@ngrx/schematics": "^12.1.0",
-    "@types/jasmine": "~3.6.0",
+    "@angular-devkit/build-angular": "~12.2.0",
+    "@angular/cli": "~12.2.0",
+    "@angular/compiler-cli": "~12.2.0",
+    "@angular/localize": "^12.2.0",
+    "@types/jasmine": "~3.8.0",
     "@types/node": "^12.11.1",
-    "codelyzer": "^6.0.0",
-    "jasmine-core": "~3.6.0",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.3",
+    "jasmine-core": "~3.8.0",
+    "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
-    "protractor": "~7.0.0",
-    "ts-node": "~8.3.0",
-    "tslint": "~6.1.0",
-    "typescript": "~4.2.4"
+    "karma-jasmine-html-reporter": "~1.7.0",
+    "typescript": "~4.3.5"
   }
 }
