@@ -1,9 +1,10 @@
+  
 import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
-import { User } from '@wiforge/averos';
 import { AverosSearchOperator, SearchInputCriteria, ApplicationSharedService } from '@wiforge/averos';
 import {Observable, of } from 'rxjs';
-import { map, mergeMap } from 'rxjs/operators';
+import { mergeMap } from 'rxjs/operators';
+
 import { ToDoArea } from '../model/to-do-area';
 
 @Injectable({
@@ -11,41 +12,42 @@ import { ToDoArea } from '../model/to-do-area';
 })
 export class ToDoAreaService {
 
-  public toDoAreaAPI = 'http://localhost:3333/api/todoareas';
+// This is an autogenerated service url
+// TODO: update with your custom service url data
 
-  private toDoTaskAPI = 'http://localhost:3333/api/todotasks';
+ public toDoAreaAPI = 'http://localhost:3333/api/todoareas';
+ public toDoTaskAPI = 'http://localhost:3333/api/todotasks';
 
+  
 
   constructor(private httpClient: HttpClient,
               private applicationSharedService: ApplicationSharedService) { }
 
-  updateEntity(id: any, toDoAreas: any | Partial<any>): Observable<any> {
-    const data = { ...toDoAreas }
+  updateEntity(id: any, entityTopdate: any | Partial<any>): Observable<any> {
+    const data = { ...entityTopdate }
     data.updatedBy =  this.applicationSharedService.getLoggedUser();
-    return this.httpClient.patch<any>(this.toDoAreaAPI + id, data);
+    data.updatedAt = new Date();
+    return this.httpClient.patch<any>(`${this.toDoAreaAPI}/${id}`, data);
   }
   
 
-  createEntity(toDoArea: any): Observable<any> {
-    toDoArea.createdBy =  this.applicationSharedService.getLoggedUser();
-    return this.httpClient.post<any>(this.toDoAreaAPI, toDoArea);
+  createEntity(value: any): Observable<any> {
+    value.createdBy =  this.applicationSharedService.getLoggedUser();
+    value.createdAt = new Date();
+    return this.httpClient.post<any>(this.toDoAreaAPI, value);
   }
 
   deleteEntity(id: any): Observable<any> {
     return this.httpClient.delete<any>(`${this.toDoAreaAPI}/${id}`);
   }
 
-  /**
-   * EndPoint: http://localhost:3333/todoarea 
-   * @returns 
-   */
   getAllEntities(): Observable<ToDoArea[]> {
     return this.httpClient.get<any[]>(this.toDoAreaAPI);
   }
 
-  getEntityById(id: string | number): Observable<any> {
+  getEntityById(id: string): Observable<any> {
 
-    return this.httpClient.get<any>(`${this.toDoAreaAPI}/${id}?_embed=toDoTasks`);
+    return this.httpClient.get<any>(`${this.toDoAreaAPI}/${id}`);
   }
 
   getEntitiesByIds(ids: string[]): Observable<any> {
@@ -57,30 +59,32 @@ export class ToDoAreaService {
                           operator: AverosSearchOperator.OPER_IN_ELEMENTS
                         }
                       };
-    return this.getEntitiesByCriteria(new SearchInputCriteria(criteriaz));
+      if (ids && ids.length > 0){
+          return this.getEntitiesByCriteria(new SearchInputCriteria(criteriaz));
+        }
+      return of();
   }
 
   getEntitiesByCriteria(criteria: SearchInputCriteria): Observable<any>{
 
-    let query: string = null;
+    let query: string;
 
     if (criteria === undefined || criteria === null){
-      return null;
+      return of();
     }   
 
     query = criteria.toHttpQuery('REGULAR');
     const opts =  {
-                    params: new HttpParams({fromString: query+'&_embed=toDoTasks'})
+                    params: new HttpParams({fromString: query})
                    };
     return this.httpClient.get<any[]>(this.toDoAreaAPI, opts);
-    // return of(Array.from(this.toDoAreas));//.filter(m => m.name === criteria.toHttpQuery()));
   }
 
   /**
    * 
-   * @param id the parent entity id (User)
-   * @param relationName : The collection relation Name (ex. "roles")
-   * @param cids : the ids of the collection to be removed from the parent entity (User) 
+   * @param id the parent entity id
+   * @param relationName : The collection relation Name
+   * @param cids : the ids of the collection to be removed from the parent entity
    * @returns 
    * 
    * N.B: cids: {id: string}[]) : cids structure depends on the identifier name of the entity:
@@ -93,40 +97,62 @@ export class ToDoAreaService {
    */
    deleteRelationCollection(parentId: any, relationName: string, cids: {id: string}[]): Observable<any> {
 
-    let operations: Observable<any> = new Observable();
-    const data: { [x: string]: User; } = null;
+    let operations: Observable<any> | null = null;
+    const data: any = {};
   
     data['updatedBy'] =  this.applicationSharedService.getLoggedUser();
+    data['updatedAt'] = new Date();
+
+    //'inline-loading' places a loading progress indicator inside the most inner html component.
+    // Set it to 'false' or ommit it in order to delegate to the global application loader located
+    // in the root html element.
     const options = {
       headers: {
         'inline-loading': 'true'
       }
     }
 
-    if (relationName === 'toDoTasks'){
-        operations = cids.reduce((previousValue: Observable<any>, currentValue: {id: string}) => {
+     /**
+       * TODO: update the target elements in the collection by removing the parent id
+       * then 
+       * update the parent's updateBy field
+       */
+
+    
+       if (relationName === 'toDoTasks'){
+        operations = cids.reduce((previousValue: Observable<any> | null, currentValue: {id: string}) => {
           previousValue = previousValue != null 
-                ? previousValue.pipe(mergeMap(el => this.httpClient.patch<any>(`${this.toDoTaskAPI}/${currentValue.id}`, {"toDoArea_id": ""}, options))) 
-                : this.httpClient.patch<any>(`${this.toDoTaskAPI}/${currentValue.id}`, {"toDoArea_id": ""}, options);
+            ? previousValue.pipe(mergeMap(el => this.httpClient.patch<any>(`${this.toDoTaskAPI}/${currentValue.id}`, {"toDoAreaId": ""}, options))) 
+            : this.httpClient.patch<any>(`${this.toDoTaskAPI}/${currentValue.id}`, {"toDoAreaId": ""}, options);
           return previousValue;
-        }, null);
+        }, operations);
+      }
+        
+if (operations !== null){
+      return  (operations as Observable<any>).pipe(mergeMap(e => this.httpClient.patch<any>(`${this.toDoAreaAPI}/${parentId}` , data)));
+    } else {
+      return this.httpClient.patch<any>(`${this.toDoAreaAPI}/${parentId}` , data);
     }
-    return operations.pipe(mergeMap(e => this.httpClient.patch<any>(`${this.toDoAreaAPI}/${parentId}` , data, options)));    
   }
 
     /**
    * 
-   * @param parentId the parent entity id (ToDoArea)
-   * @param relationName : The collection relation Name (ex. "roles")
-   * @param cids : the ids of the collection to be added from the parent entity (User) 
+   * @param parentId the parent entity id 
+   * @param relationName : The collection relation Name
+   * @param cids : the ids of the collection to be added from the parent entity
    * @returns 
    */
      addRelationCollection(parentId: any, relationName: string, cids: {id: string}[]): Observable<any> {
 
-      let operations: Observable<any> = new Observable();
-      const data: { [x: string]: User; } = null;
+      let operations: Observable<any> | null = null;
+      const data: any = {};
     
       data['updatedBy'] =  this.applicationSharedService.getLoggedUser();
+      data['updatedAt'] = new Date();
+
+      //'inline-loading' places a loading progress indicator inside the most inner html component.
+      // Set it to 'false' or ommit it in order to delegate to the global application loader located
+      // in the root html element.
       const options = {
         headers: {
           'inline-loading': 'true'
@@ -134,25 +160,26 @@ export class ToDoAreaService {
       }
 
       /**
-       * update the target elements in the collection by setting the parent id
+       * TODO: update the target elements in the collection by setting the parent id
        * then 
        * update the parent's updateBy field
        */
-      if (relationName === 'toDoTasks'){
-      // Handles the relation 'toDoTasks': Updates all the childs in a seperate chained requests (mergeMap()),
-      //   then updates the parent entity (the owner) with the latest 'updateBy' user
 
-      // Could also use forkJoin for parallel update on the collection
-        operations = cids.reduce((previousValue: Observable<any>, currentValue: {id: string}) => {
+      
+       if (relationName === 'toDoTasks'){
+        operations = cids.reduce((previousValue: Observable<any> | null, currentValue: {id: string}) => {
           previousValue = previousValue != null 
-                ? previousValue.pipe(mergeMap(el => this.httpClient.patch<any>(`${this.toDoTaskAPI}/${currentValue.id}`, {"toDoArea_id": parentId}))) 
-                : this.httpClient.patch<any>(`${this.toDoTaskAPI}/${currentValue.id}`, {"toDoArea_id": parentId});
+            ? previousValue.pipe(mergeMap(el => this.httpClient.patch<any>(`${this.toDoTaskAPI}/${currentValue.id}`, {"toDoAreaId": parentId}, options))) 
+            : this.httpClient.patch<any>(`${this.toDoTaskAPI}/${currentValue.id}`, {"toDoAreaId": parentId}, options);
           return previousValue;
-        }, null);
+        }, operations);
       }
-
-        return operations.pipe(mergeMap(e => this.httpClient.patch<any>(`${this.toDoAreaAPI}/${parentId}` , data)));
+        
+if (operations !== null){
+          return  (operations as Observable<any>).pipe(mergeMap(e => this.httpClient.patch<any>(`${this.toDoAreaAPI}/${parentId}` , data)));
+       } else {
+          return this.httpClient.patch<any>(`${this.toDoAreaAPI}/${parentId}` , data);
+       }
     }
-
-
+  
 }
