@@ -1,3 +1,4 @@
+
 import { Injectable } from '@angular/core';
 import { ViewLayoutService } from '@wiforge/averos';
 import { ToDoArea } from '../model/to-do-area';
@@ -8,17 +9,14 @@ import { ToDoTask } from '../model/to-do-task';
 })
 export class ApplicationInitializerService {
 
-  private registeredEntities: Array<any> = [ToDoTask.instanceMetadata, ToDoArea.instanceMetadata];
+  private registeredEntities: Array<any> = [ ToDoArea.instanceMetadata, ToDoTask.instanceMetadata];
 
   constructor(private viewLayoutService: ViewLayoutService) { }
 
 
   initialize(): Promise<any> {
-    const p1 = new Promise ((resolve, reject) => {
+    return Promise.all([new Promise ((resolve, reject) => {
       this.viewLayoutService.registerEntitiesViewLayouts(this.registeredEntities);
-    resolve(true)});
-    const asyncInitPromises: Promise<any>[] = [p1];
- 
-    return Promise.all(asyncInitPromises);// Wait for all promises to execute
+      resolve(true)})]);
   }
 }
