@@ -2,29 +2,28 @@ import { ChangeDetectionStrategy, Component, OnDestroy, OnInit, ViewChild } from
 import { FormGroup } from '@angular/forms';
 import { MatDialog, MatDialogConfig } from '@angular/material/dialog';
 import { ActivatedRoute, NavigationExtras, Params, Router } from '@angular/router';
-import { AlertService, AverosDialogViewConfig, AverosDynamicDialogComponent,
-         AverosSearchEntityComponent, AverosViewConfig, FormControlService, 
-         SearchInputCriteria, SearchUseCase, TypeScriptTypeMetaDatatHandler, 
-         UseCase, UseCaseConfig, User, ViewLayoutService } from '@wiforge/averos';
+import { SearchInputCriteria, AlertService, FormControlService, ViewLayoutService,
+         TypeScriptTypeMetaDatatHandler, AverosDynamicDialogComponent, AverosSearchEntityComponent, 
+         UseCase, SearchUseCase, UseCaseConfig, AverosDialogViewConfig, AverosViewConfig, User
+        } from '@wiforge/averos';
 import { Observable, of, Subscription } from 'rxjs';
 import { map } from 'rxjs/operators';
 import { ToDoArea } from '../../../model/to-do-area';
-import { ToDoTask } from '../../../model/to-do-task';
 import { ToDoAreaService } from '../../../service/to-do-area.service';
+import { ToDoTask } from '../../../model/to-do-task';
 
 
 @Component({
-  selector: 'app-search-to-do-area',
-  templateUrl: './search-to-do-area.component.html',
-  styleUrls: ['./search-to-do-area.component.scss'],
-  changeDetection: ChangeDetectionStrategy.OnPush
+	selector: 'app-search-to-do-area-component',
+	templateUrl: './search-to-do-area.component.html',
+	styleUrls: ['./search-to-do-area.component.scss'],
+    changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class SearchToDoAreaComponent implements SearchUseCase<ToDoArea>, OnInit, OnDestroy {
 
-  @ViewChild(AverosSearchEntityComponent, {static: true}) averosSearchEntity: AverosSearchEntityComponent<ToDoArea>;
-
-
- targetEntity =  ToDoArea.instanceMetadata;
+@ViewChild(AverosSearchEntityComponent, {static: true}) averosSearchEntity!: AverosSearchEntityComponent<ToDoArea>;
+  
+  targetEntity =  ToDoArea.instanceMetadata;
  viewLayout = ToDoArea.getEntityViewLayout();
  searchInputFormGoup: FormGroup;// = this.buildForm();
 
@@ -33,13 +32,13 @@ export class SearchToDoAreaComponent implements SearchUseCase<ToDoArea>, OnInit,
                                                   componentAppearance: 'outline',
                                                   iconLayout: 'component',
                                                   entityType: ToDoArea,
-                                                  entity: null
+                                                  entity: undefined
                                                 };
   showSearchResult = false;
-  searchCriteria: SearchInputCriteria;
-  tableValues$: Observable<ToDoArea[]> ;
+  searchCriteria!: SearchInputCriteria;
+  tableValues$!: Observable<ToDoArea[]> ;
   private genericDialog = new MatDialogConfig();
-  private deleteSubscription: Subscription;
+  private deleteSubscription!: Subscription;
 
 
   /**
@@ -56,9 +55,9 @@ export class SearchToDoAreaComponent implements SearchUseCase<ToDoArea>, OnInit,
 
     } else if (avConfig.useCaseConfig?.useCase===UseCase.EDIT ||
                 avConfig.useCaseConfig?.useCase===UseCase.UPDATE ){
-     return this.entityService.updateEntity(avConfig.useCaseConfig.entity?.id, objectSubjectToAction)
+     return this.entityService.updateEntity(avConfig.value[TypeScriptTypeMetaDatatHandler.instance.getIdName(this.targetEntity)], objectSubjectToAction);
     }
-    return null;
+    return of();
     
   }
   constructor(private entityService: ToDoAreaService,
@@ -72,6 +71,7 @@ export class SearchToDoAreaComponent implements SearchUseCase<ToDoArea>, OnInit,
                }
 
   ngOnDestroy(): void {
+
     this.deleteSubscription?.unsubscribe();
   }
   
@@ -85,7 +85,7 @@ export class SearchToDoAreaComponent implements SearchUseCase<ToDoArea>, OnInit,
         usecase: UseCase.EDIT,
       }
     };
-    this.router.navigate(['/todoarea/edit', entity[TypeScriptTypeMetaDatatHandler.instance.getIdName(this.targetEntity)]], navigationExtras);
+    this.router.navigate(['/todoareas/edit', entity[TypeScriptTypeMetaDatatHandler.instance.getIdName(this.targetEntity)]], navigationExtras);
   }
   delete(valueToBeDeleted: ToDoArea) {
     const idName = TypeScriptTypeMetaDatatHandler.instance.getIdName(this.targetEntity);
@@ -94,7 +94,6 @@ export class SearchToDoAreaComponent implements SearchUseCase<ToDoArea>, OnInit,
               (deletedEntity: ToDoArea) => {
                 this.alertService.success($localize`:@@uc.delete.entity:Entity ${valueToBeDeleted.name}:entity:
                 has been deleted successfully`);
-                // eslint-disable-next-line no-underscore-dangle
                 this.tableValues$ = this.tableValues$.pipe(map(e=>e.filter(el=> (el as any)[idName] !== (valueToBeDeleted as any)[idName])));
                 this.deleteSubscription?.unsubscribe();
               },
@@ -117,10 +116,11 @@ export class SearchToDoAreaComponent implements SearchUseCase<ToDoArea>, OnInit,
                               objectClass: ToDoArea, 
                               compositeObject: {value: ToDoArea.instanceMetadata, type: 'ToDoArea'},
                               onSubmitCallback: this.onAddOrUpateCallback,
-                              onLoadCallback: null,
+                              onLoadCallback: undefined,
                               viewLayout : { 
                                             useCase: UseCase.CREATE,
-                                            editMode: true
+                                            editMode: true,
+                                            canActivateEditMode: true
                                           }
                             };
 
@@ -136,9 +136,13 @@ export class SearchToDoAreaComponent implements SearchUseCase<ToDoArea>, OnInit,
         usecase: UseCase.VIEW,
       }
     };
-    this.router.navigate(['/todoarea/view/', entity[TypeScriptTypeMetaDatatHandler.instance.getIdName(this.targetEntity)]], navigationExtras);
+    this.router.navigate(['/todoareas/view/', entity[TypeScriptTypeMetaDatatHandler.instance.getIdName(this.targetEntity)]], navigationExtras);
   }
 
+/**
+* This function handles a one-to-one and one-to-many relationships in a view composite relation use case
+* Please do not modify or update the function structure since it's is auto-updated when including additional entity members
+*/
   viewCompositeObject(compositeObject: { value: unknown; type: string; compositeType?: string; }) {
     let averosDialogViewConfig: AverosDialogViewConfig =
     
@@ -146,24 +150,28 @@ export class SearchToDoAreaComponent implements SearchUseCase<ToDoArea>, OnInit,
                                   objectClass: null, ////example User (type not string)
                                   compositeObject: compositeObject,
                                   onSubmitCallback: (a: any,b: any)=>{},
-                                  onLoadCallback:  null,
-                                  viewLayout : { useCase: null,
+                                  onLoadCallback:  undefined,
+                                  viewLayout : { useCase: -1,
                                                  editMode: false,
                                                  canActivateEditMode: false
                                                }
         };
 
     if (compositeObject?.compositeType === 'collection'){
-          if (compositeObject.type === 'ToDoTask'){// composite of type Collection (list of assigned ToDoTasks)
-            averosDialogViewConfig.objectClass= ToDoTask;    
-            averosDialogViewConfig.viewLayout.useCase = UseCase.SEARCH_RESULT_TABLE;    
-            averosDialogViewConfig.onLoadCallback = (entities: any[]): Observable<any> => of(entities);                
-          }
+      if (compositeObject.type === 'ToDoTask'){
+      averosDialogViewConfig.objectClass= ToDoTask;
+      averosDialogViewConfig.viewLayout.useCase = UseCase.SEARCH_RESULT_TABLE;    
+      averosDialogViewConfig.onLoadCallback = (entities: any[]): Observable<any> => of(entities); 
+    }
+      
+ /// handles One-To-Many relationship if any
+
     }
     else {
       averosDialogViewConfig.viewLayout.useCase = UseCase.VIEW;
-      if (compositeObject.type === 'User'){// User Entity
+      if (compositeObject.type === 'User'){// User Entity this is an Averos type but it could be overriden at your convenience
         averosDialogViewConfig.objectClass= User;
+
       }
     }
 
