@@ -1,38 +1,42 @@
 import { ChangeDetectorRef, OnDestroy } from '@angular/core';
 import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
-import { Location } from '@angular/common';
 import { ActivatedRoute, Router } from '@angular/router';
+import { Location } from '@angular/common';
 import { FormGroup } from '@angular/forms';
 import { Observable, Subscription } from 'rxjs';
+import { slideInOutAnimation,CreateViewEditUseCase, UseCase, UseCaseConfig,
+		 AlertService, FormControlService, EntityAlteredRelationEventData, TypeScriptTypeMetaDatatHandler,
+		 UseCaseAction, UseCaseViewLayout } from '@wiforge/averos';
 import { ToDoArea } from '../../../model/to-do-area';
 import { ToDoAreaService } from '../../../service/to-do-area.service';
-import { ToDoTask } from '../../../model/to-do-task';
-import { AlertService, CreateViewEditUseCase, EntityAlteredRelationEventData, 
-         FormControlService, slideInOutAnimation, TypeScriptTypeMetaDatatHandler, UseCase, 
-         UseCaseAction, UseCaseConfig, UseCaseViewLayout } from '@wiforge/averos';
+
 
 @Component({
-  selector: 'app-to-do-area-details',
+  selector: 'app-to-do-area-details-component',
   templateUrl: './to-do-area-details.component.html',
   styleUrls: ['./to-do-area-details.component.scss'],
   animations: [slideInOutAnimation],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ToDoAreaDetailsComponent implements CreateViewEditUseCase<ToDoArea>, OnInit, OnDestroy {
-  useCaseViewLayout: Observable<UseCaseViewLayout<ToDoArea>> = null; 
-  reactiveForm: FormGroup = null;
+  useCaseViewLayout!: Observable<UseCaseViewLayout<ToDoArea>>; 
+  reactiveForm!: FormGroup;
   useCaseConfig: UseCaseConfig<ToDoArea> = {
                                           componentAppearance: 'outline',
                                           iconLayout: 'component',
-                                          entity: null,
+                                          entity: undefined,
                                           entityType: ToDoArea,
-                                          useCase: null
+                                          useCase: undefined
                                        };
   editModeActivated = false;
   currentUseCase: UseCase;
 
-  useCaseEntity: ToDoArea;
-  userSubscription: Subscription;
+  useCaseEntity!: ToDoArea;
+  private userSubscription!: Subscription;
+  private addRelationSubscription!: Subscription;
+  private deleteRelationSubscription!: Subscription;
+  private updateEntitySubscription!: Subscription;
+  private createEntitySubscription!: Subscription;
 
   constructor(private entityService: ToDoAreaService,
               private alertService: AlertService,
@@ -43,7 +47,7 @@ export class ToDoAreaDetailsComponent implements CreateViewEditUseCase<ToDoArea>
               private cd: ChangeDetectorRef) {
 
     const navigation = this.router.getCurrentNavigation();
-    this.currentUseCase = navigation.extras.state?.usecase ? navigation.extras.state?.usecase : UseCase.VIEW;
+    this.currentUseCase = navigation?.extras.state?.usecase ? navigation.extras.state?.usecase : UseCase.VIEW;
 
     if (this.currentUseCase === UseCase.EDIT || 
         this.currentUseCase === UseCase.CREATE || 
@@ -59,6 +63,10 @@ export class ToDoAreaDetailsComponent implements CreateViewEditUseCase<ToDoArea>
 
   ngOnDestroy(): void {
     this.userSubscription?.unsubscribe();
+    this.addRelationSubscription?.unsubscribe();
+    this.deleteRelationSubscription?.unsubscribe();
+    this.updateEntitySubscription?.unsubscribe();
+    this.createEntitySubscription?.unsubscribe();
   }
 
   get useCase() {
@@ -68,29 +76,32 @@ export class ToDoAreaDetailsComponent implements CreateViewEditUseCase<ToDoArea>
   onSubmit(submittedValue: ToDoArea) {
     // handles updates in case of EDIT use case
     if (this.currentUseCase === UseCase.EDIT || this.currentUseCase === UseCase.UPDATE){
-      this.entityService.updateEntity(this.useCaseConfig.entity?.id, submittedValue).subscribe(
+       this.updateEntitySubscription = this.entityService.updateEntity(this.useCaseConfig.entity?.id, submittedValue).subscribe(
         (updatedEntity: ToDoArea) => {
           this.alertService.success($localize`:@@uc.update.entity:Entity ${updatedEntity.name}:entity:
            has been updated successfully`);
           
           this.editModeActivated = false;
           this.updateView(UseCase.VIEW, updatedEntity);
+          this.updateEntitySubscription?.unsubscribe();
         },
         err => {
           console.log(err);
+          this.updateEntitySubscription?.unsubscribe();
         });
     } else if (this.currentUseCase === UseCase.CREATE){
-      this.entityService.createEntity(submittedValue).subscribe(
-        (updatedEntity: ToDoArea) => {
-          this.alertService.success($localize`:@@uc.create.entity:Entity ${updatedEntity.name}:entity:
-           has been updated successfully`);
+      this.createEntitySubscription = this.entityService.createEntity(submittedValue).subscribe(
+        (createdEntity: ToDoArea) => {
+          this.alertService.success($localize`:@@uc.create.entity:Entity ${createdEntity.name}:entity:
+           has been created successfully`);
           this.editModeActivated = false;
-          this.updateView(UseCase.VIEW, updatedEntity);
+          this.updateView(UseCase.VIEW, createdEntity);
+          this.createEntitySubscription?.unsubscribe();
          
         },
         err => {
           console.log(err);
-          // this.alertService.error(err);
+          this.createEntitySubscription?.unsubscribe();
         });
     }
 
@@ -101,42 +112,41 @@ export class ToDoAreaDetailsComponent implements CreateViewEditUseCase<ToDoArea>
    * either by adding a new value (in case of addition) or by deleting an existing value from the related collection
    */
   updateRelationCollection(entityAlteredRelationEventData: EntityAlteredRelationEventData){
-  const memberType = TypeScriptTypeMetaDatatHandler.instance.getMemberType(ToDoArea.instanceMetadata, entityAlteredRelationEventData.actionEventData.relationName);
-  /// Handles relation with Role
-  if (memberType instanceof ToDoTask){
+     let  idName = TypeScriptTypeMetaDatatHandler.instance.getIdName(this.useCaseConfig.entityType)
     if (entityAlteredRelationEventData.actionEventData.formattedIdsSubjectToAction.length > 0){
-      
       // handle the relation collection data by action
       switch(entityAlteredRelationEventData.action) {
         case UseCaseAction.DELETE:
-              /// delete an element from the list of roles that are assigned to the user
-              this.entityService.deleteRelationCollection(this.useCaseEntity.id, 
+              this.deleteRelationSubscription = this.entityService.deleteRelationCollection((this.useCaseEntity as any)[idName], 
                 entityAlteredRelationEventData.actionEventData.relationName,
                 entityAlteredRelationEventData.actionEventData.formattedIdsSubjectToAction).subscribe(
                   (updates: any) => {
                     this.alertService.success($localize`:@@uc.update.entity:Entity ${this.useCaseEntity.name}:entity:
-                      has been updated successfully`);                                                     
+                      has been updated successfully`); 
+                    this.deleteRelationSubscription?.unsubscribe();                                                     
                   },
                   err => {
                     console.log(err);
                     this.alertService.error($localize`:@@uc.update.entity.error:Entity ${this.useCaseEntity.name}:entity:
-                      cannot be updated`);                                                     
+                      cannot be updated`); 
+                    this.deleteRelationSubscription?.unsubscribe();                                                    
                   }
                 );
           break;
         case UseCaseAction.ADD:
-              /// add elements from to the list of roles that are assigned to the user
-              this.entityService.addRelationCollection(this.useCaseEntity.id, 
+              this.addRelationSubscription = this.entityService.addRelationCollection((this.useCaseEntity as any)[idName], 
                 entityAlteredRelationEventData.actionEventData.relationName,
                 entityAlteredRelationEventData.actionEventData.formattedIdsSubjectToAction).subscribe(
                   (updates: any) => {
                     this.alertService.success($localize`:@@uc.update.entity:Entity ${this.useCaseEntity.name}:entity:
-                      has been updated successfully`);                                                     
+                      has been updated successfully`);        
+                    this.addRelationSubscription?.unsubscribe();                                              
                   },
                   err => {
                     console.log(err);
                     this.alertService.error($localize`:@@uc.update.entity.error:Entity ${this.useCaseEntity.name}:entity:
-                      cannot be updated`);                                                     
+                      cannot be updated`);  
+                    this.addRelationSubscription?.unsubscribe();                                                    
                   }
                 );
             
@@ -144,10 +154,7 @@ export class ToDoAreaDetailsComponent implements CreateViewEditUseCase<ToDoArea>
         default:
           break;
       }
-    }    
-  }
-
-
+    }  
   }
 
   ngOnInit(): void {
@@ -174,13 +181,14 @@ export class ToDoAreaDetailsComponent implements CreateViewEditUseCase<ToDoArea>
    */
      updateUseCaseViewData(useCase: UseCase) {
 
-      // const id = +this.route.snapshot.paramMap.get('id');
       const id = this.route.snapshot.paramMap.get('id');
   
-      this.userSubscription = this.entityService.getEntityById(id)
-        .subscribe(entity => {
-          this.updateView(useCase, entity);
-        });
+      if (id){
+        this.userSubscription = this.entityService.getEntityById(id)
+          .subscribe(entity => {
+            this.updateView(useCase, entity);
+          });
+      }
     }
 
     /**
