@@ -1,34 +1,42 @@
 import { ChangeDetectorRef, OnDestroy } from '@angular/core';
 import { Component, OnInit, ChangeDetectionStrategy } from '@angular/core';
+import { ActivatedRoute, Router } from '@angular/router';
 import { Location } from '@angular/common';
 import { FormGroup } from '@angular/forms';
 import { Observable, Subscription } from 'rxjs';
-import { ActivatedRoute, Router } from '@angular/router';
+import { slideInOutAnimation,CreateViewEditUseCase, UseCase, UseCaseConfig,
+		 AlertService, FormControlService, EntityAlteredRelationEventData, TypeScriptTypeMetaDatatHandler,
+		 UseCaseAction, UseCaseViewLayout } from '@wiforge/averos';
 import { ToDoTask } from '../../../model/to-do-task';
 import { ToDoTaskService } from '../../../service/to-do-task.service';
-import { AlertService, CreateViewEditUseCase, EntityAlteredRelationEventData, FormControlService, slideInOutAnimation, TypeScriptTypeMetaDatatHandler, UseCase, UseCaseAction, UseCaseConfig, UseCaseViewLayout } from '@wiforge/averos';
+
+
 @Component({
-  selector: 'app-to-do-task-details',
+  selector: 'app-to-do-task-details-component',
   templateUrl: './to-do-task-details.component.html',
   styleUrls: ['./to-do-task-details.component.scss'],
   animations: [slideInOutAnimation],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class ToDoTaskDetailsComponent implements CreateViewEditUseCase<ToDoTask>, OnInit, OnDestroy {
-  useCaseViewLayout: Observable<UseCaseViewLayout<ToDoTask>> = null; 
-  reactiveForm: FormGroup = null;
+  useCaseViewLayout!: Observable<UseCaseViewLayout<ToDoTask>>; 
+  reactiveForm!: FormGroup;
   useCaseConfig: UseCaseConfig<ToDoTask> = {
                                           componentAppearance: 'outline',
                                           iconLayout: 'component',
-                                          entity: null,
+                                          entity: undefined,
                                           entityType: ToDoTask,
-                                          useCase: null
+                                          useCase: undefined
                                        };
   editModeActivated = false;
   currentUseCase: UseCase;
 
-  useCaseEntity: ToDoTask;
-  userSubscription: Subscription;
+  useCaseEntity!: ToDoTask;
+  private userSubscription!: Subscription;
+  private addRelationSubscription!: Subscription;
+  private deleteRelationSubscription!: Subscription;
+  private updateEntitySubscription!: Subscription;
+  private createEntitySubscription!: Subscription;
 
   constructor(private entityService: ToDoTaskService,
               private alertService: AlertService,
@@ -39,7 +47,7 @@ export class ToDoTaskDetailsComponent implements CreateViewEditUseCase<ToDoTask>
               private cd: ChangeDetectorRef) {
 
     const navigation = this.router.getCurrentNavigation();
-    this.currentUseCase = navigation.extras.state?.usecase ? navigation.extras.state?.usecase : UseCase.VIEW;
+    this.currentUseCase = navigation?.extras.state?.usecase ? navigation.extras.state?.usecase : UseCase.VIEW;
 
     if (this.currentUseCase === UseCase.EDIT || 
         this.currentUseCase === UseCase.CREATE || 
@@ -53,96 +61,100 @@ export class ToDoTaskDetailsComponent implements CreateViewEditUseCase<ToDoTask>
     
   }
 
-  /**
-   * updateRelationCollection updates the related entity with the resulted collection
-   * either by adding a new value (in case of addition) or by deleting an existing value from the related collection
-   */
-   updateRelationCollection(entityAlteredRelationEventData: EntityAlteredRelationEventData){
-    const memberType = TypeScriptTypeMetaDatatHandler.instance.getMemberType(ToDoTask.instanceMetadata, entityAlteredRelationEventData.actionEventData.relationName);
-    /// Handles relation with Role
-    if (memberType instanceof ToDoTask){
-      if (entityAlteredRelationEventData.actionEventData.formattedIdsSubjectToAction.length > 0){
-        
-        // handle the relation collection data by action
-        switch(entityAlteredRelationEventData.action) {
-          case UseCaseAction.DELETE:
-                /// delete an element from the list of roles that are assigned to the user
-                this.entityService.deleteRelationCollection(this.useCaseEntity.id, 
-                  entityAlteredRelationEventData.actionEventData.relationName,
-                  entityAlteredRelationEventData.actionEventData.formattedIdsSubjectToAction).subscribe(
-                    (updates: any) => {
-                      this.alertService.success($localize`:@@uc.update.entity:Entity ${this.useCaseEntity.name}:entity:
-                        has been updated successfully`);                                                     
-                    },
-                    err => {
-                      console.log(err);
-                      this.alertService.error($localize`:@@uc.update.entity.error:Entity ${this.useCaseEntity.name}:entity:
-                        cannot be updated`);                                                     
-                    }
-                  );
-            break;
-          case UseCaseAction.ADD:
-                /// add elements from to the list of roles that are assigned to the user
-                this.entityService.addRelationCollection(this.useCaseEntity.id, 
-                  entityAlteredRelationEventData.actionEventData.relationName,
-                  entityAlteredRelationEventData.actionEventData.formattedIdsSubjectToAction).subscribe(
-                    (updates: any) => {
-                      this.alertService.success($localize`:@@uc.update.entity:User ${this.useCaseEntity.name}:entity:
-                        has been updated successfully`);                                                     
-                    },
-                    err => {
-                      console.log(err);
-                      this.alertService.error($localize`:@@uc.update.entity.error:User ${this.useCaseEntity.name}:entity:
-                        cannot be updated`);                                                     
-                    }
-                  );
-              
-            break;
-          default:
-            break;
-        }
-      }    
-    }
-  
-  
-    }
-
   ngOnDestroy(): void {
     this.userSubscription?.unsubscribe();
+    this.addRelationSubscription?.unsubscribe();
+    this.deleteRelationSubscription?.unsubscribe();
+    this.updateEntitySubscription?.unsubscribe();
+    this.createEntitySubscription?.unsubscribe();
   }
 
   get useCase() {
     return UseCase;
   }
   
-  onSubmit(submittedValue: ToDoTask) { 
+  onSubmit(submittedValue: ToDoTask) {
     // handles updates in case of EDIT use case
     if (this.currentUseCase === UseCase.EDIT || this.currentUseCase === UseCase.UPDATE){
-      this.entityService.updateEntity(this.useCaseConfig.entity?.id, submittedValue).subscribe(
+       this.updateEntitySubscription = this.entityService.updateEntity(this.useCaseConfig.entity?.id, submittedValue).subscribe(
         (updatedEntity: ToDoTask) => {
           this.alertService.success($localize`:@@uc.update.entity:Entity ${updatedEntity.name}:entity:
            has been updated successfully`);
           
           this.editModeActivated = false;
           this.updateView(UseCase.VIEW, updatedEntity);
+          this.updateEntitySubscription?.unsubscribe();
         },
         err => {
           console.log(err);
+          this.updateEntitySubscription?.unsubscribe();
         });
     } else if (this.currentUseCase === UseCase.CREATE){
-      this.entityService.createEntity(submittedValue).subscribe(
-        (updatedEntity: ToDoTask) => {
-          this.alertService.success($localize`:@@uc.create.entity:Entity ${updatedEntity.name}:entity:
-           has been updated successfully`);
+      this.createEntitySubscription = this.entityService.createEntity(submittedValue).subscribe(
+        (createdEntity: ToDoTask) => {
+          this.alertService.success($localize`:@@uc.create.entity:Entity ${createdEntity.name}:entity:
+           has been created successfully`);
           this.editModeActivated = false;
-          this.updateView(UseCase.VIEW, updatedEntity);
+          this.updateView(UseCase.VIEW, createdEntity);
+          this.createEntitySubscription?.unsubscribe();
          
         },
         err => {
           console.log(err);
+          this.createEntitySubscription?.unsubscribe();
         });
     }
 
+  }
+
+   /**
+   * updateRelationCollection updates the related entity with the resulted collection
+   * either by adding a new value (in case of addition) or by deleting an existing value from the related collection
+   */
+  updateRelationCollection(entityAlteredRelationEventData: EntityAlteredRelationEventData){
+     let  idName = TypeScriptTypeMetaDatatHandler.instance.getIdName(this.useCaseConfig.entityType)
+    if (entityAlteredRelationEventData.actionEventData.formattedIdsSubjectToAction.length > 0){
+      // handle the relation collection data by action
+      switch(entityAlteredRelationEventData.action) {
+        case UseCaseAction.DELETE:
+              this.deleteRelationSubscription = this.entityService.deleteRelationCollection((this.useCaseEntity as any)[idName], 
+                entityAlteredRelationEventData.actionEventData.relationName,
+                entityAlteredRelationEventData.actionEventData.formattedIdsSubjectToAction).subscribe(
+                  (updates: any) => {
+                    this.alertService.success($localize`:@@uc.update.entity:Entity ${this.useCaseEntity.name}:entity:
+                      has been updated successfully`); 
+                    this.deleteRelationSubscription?.unsubscribe();                                                     
+                  },
+                  err => {
+                    console.log(err);
+                    this.alertService.error($localize`:@@uc.update.entity.error:Entity ${this.useCaseEntity.name}:entity:
+                      cannot be updated`); 
+                    this.deleteRelationSubscription?.unsubscribe();                                                    
+                  }
+                );
+          break;
+        case UseCaseAction.ADD:
+              this.addRelationSubscription = this.entityService.addRelationCollection((this.useCaseEntity as any)[idName], 
+                entityAlteredRelationEventData.actionEventData.relationName,
+                entityAlteredRelationEventData.actionEventData.formattedIdsSubjectToAction).subscribe(
+                  (updates: any) => {
+                    this.alertService.success($localize`:@@uc.update.entity:Entity ${this.useCaseEntity.name}:entity:
+                      has been updated successfully`);        
+                    this.addRelationSubscription?.unsubscribe();                                              
+                  },
+                  err => {
+                    console.log(err);
+                    this.alertService.error($localize`:@@uc.update.entity.error:Entity ${this.useCaseEntity.name}:entity:
+                      cannot be updated`);  
+                    this.addRelationSubscription?.unsubscribe();                                                    
+                  }
+                );
+            
+          break;
+        default:
+          break;
+      }
+    }  
   }
 
   ngOnInit(): void {
@@ -169,13 +181,14 @@ export class ToDoTaskDetailsComponent implements CreateViewEditUseCase<ToDoTask>
    */
      updateUseCaseViewData(useCase: UseCase) {
 
-      // const id = +this.route.snapshot.paramMap.get('id');
       const id = this.route.snapshot.paramMap.get('id');
   
-      this.userSubscription = this.entityService.getEntityById(id)
-        .subscribe(entity => {
-          this.updateView(useCase, entity);
-        });
+      if (id){
+        this.userSubscription = this.entityService.getEntityById(id)
+          .subscribe(entity => {
+            this.updateView(useCase, entity);
+          });
+      }
     }
 
     /**
@@ -202,4 +215,3 @@ export class ToDoTaskDetailsComponent implements CreateViewEditUseCase<ToDoTask>
   }
 
 }
-
