@@ -1,6 +1,4 @@
-/***************************************************************************************************
- * Load `$localize` onto the global scope - used if i18n tags appear in Angular templates.
- */
+
 import '@angular/localize/init';
 /**
  * This file includes polyfills needed by Angular and is loaded before the app.
